example-project/pyproject.toml
[tool.poetry]
name = "sf-demo"
version = "0.1.0"
description = "example yelp.com scraper using scrapfly SDK"
authors = ["granitosaurus <bernardas.alisauskas@pm.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
scrapfly-sdk = {extras = ["all"], version = "^0.8.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
